+++ packages/gasket-command-plugin/test/plugin.test.js
@@ -1,49 +1,77 @@
-const plugin = require('..');
+const assume = require('assume');
+const sinon = require('sinon');
+const plugin = require('../lib/plugin');
+const GasketCommand = require('../lib/command');
 
-class MockBaseCommand {}
 
 describe('The plugin', () => {
-  let gasket;
+  let gasket, execStub, oclifConfig, MockCmdA, MockCmdB, MockCmdC;
 
   beforeEach(() => {
+    execStub = sinon.stub();
     gasket = {
-      exec: jest.fn()
+      exec: execStub
     };
+    oclifConfig = { plugins: [] };
+
+    MockCmdA = class extends GasketCommand {};
+    MockCmdA.id = 'a';
+    MockCmdB = class extends GasketCommand {};
+    MockCmdB.id = 'b';
+    MockCmdC = class extends GasketCommand {};
+    MockCmdC.id = 'c';
   });
 
   describe('initOclif hook', () => {
-    it('injects a plugin with commands from getCommands hooks', async () => {
-      const oclifConfig = { plugins: [] };
-      gasket.exec.mockImplementation(async (event, args) => {
-        if (event === 'getCommands') {
-          class SomeCommand extends args.BaseCommand {}
-          SomeCommand.id = 'new';
-          return [SomeCommand];
-        }
-      });
 
-      await plugin.hooks.initOclif(gasket, {
-        oclifConfig,
-        BaseCommand: MockBaseCommand
-      });
+    it('executes getCommands lifecycle', async () => {
+      execStub.resolves([]);
+      await plugin.hooks.initOclif(gasket, { oclifConfig });
+      assume(execStub).calledWith('getCommands');
+    });
 
-      expect(gasket.exec).toHaveBeenCalledWith('getCommands', {
-        oclifConfig,
-        BaseCommand: MockBaseCommand
-      });
+    it('passes GasketCommand to getCommands lifecycle args', async () => {
+      execStub.resolves([]);
+      await plugin.hooks.initOclif(gasket, { oclifConfig });
+      const results = execStub.getCall(0).args[1];
+      assume(results).property('GasketCommand', GasketCommand);
+    });
 
-      const injectedPlugin = oclifConfig.plugins[0];
-      expect(injectedPlugin.commands).toHaveLength(1);
+    it('passes oclif flags to getCommands lifecycle args', async () => {
+      execStub.resolves([]);
+      await plugin.hooks.initOclif(gasket, { oclifConfig });
+      const results = execStub.getCall(0).args[1];
+      assume(results).property('flags', require('@oclif/command').flags);
+    });
 
-      const injectedCommandMeta = injectedPlugin.commands[0];
-      expect(injectedCommandMeta).toHaveProperty('id', 'new');
-      expect(injectedCommandMeta).toHaveProperty('load');
+    it('injects plugin with commands to oclifConfig', async () => {
+      execStub.resolves([MockCmdA]);
+      await plugin.hooks.initOclif(gasket, { oclifConfig });
+      assume(oclifConfig.plugins).lengthOf(1);
+      assume(oclifConfig.plugins[0].name).equals('Gasket commands');
+      assume(oclifConfig.plugins[0].commands[0].id).equals(MockCmdA.id);
+    });
 
-      const Command = await injectedCommandMeta.load();
-      expect(typeof Command).toEqual('function');
+    it('flattens and filters undefined for getCommands results', async () => {
+      // eslint-disable-next-line no-undefined
+      execStub.resolves([[MockCmdA, undefined], MockCmdB, [undefined, MockCmdC], undefined]);
+      await plugin.hooks.initOclif(gasket, { oclifConfig });
+      assume(oclifConfig.plugins[0].commands).lengthOf(3);
+    });
 
-      const command = new Command();
-      expect(command).toBeInstanceOf(MockBaseCommand);
+    it('hoists flags from base GasketCommand', async () => {
+      execStub.resolves([MockCmdA]);
+      await plugin.hooks.initOclif(gasket, { oclifConfig });
+      assume(MockCmdA).property('flags');
+      assume(MockCmdA.flags).eqls(GasketCommand.flags);
+    });
+
+    it('adds load method to injected Command', async () => {
+      execStub.resolves([MockCmdA]);
+      await plugin.hooks.initOclif(gasket, { oclifConfig });
+      const injectedCmd = oclifConfig.plugins[0].commands[0];
+      assume(injectedCmd).property('load');
+      assume(injectedCmd.load).instanceOf(Function);
     });
   });
 });